♻️🚚: move echo api tests into package
